requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
frozenlist==1.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
multidict==6.0.4
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
yarl==1.9.2
